changed lookback_params, debugging, added helper functions
