+++ src/modelfree/common/transparent.py
@@ -4,7 +4,6 @@ import numpy as np
 from stable_baselines.common.policies import FeedForwardPolicy, nature_cnn
 import tensorflow as tf
 
-from aprl.envs.multi_agent import CurryVecEnv, _tuple_pop
 from modelfree.common.utils import _filter_dict
 
 TRANSPARENCY_KEYS = set(['obs', 'ff_policy', 'ff_value', 'hid'])
@@ -67,34 +66,3 @@ class TransparentMlpPolicy(TransparentFeedForwardPolicy):
         super(TransparentMlpPolicy, self).__init__(sess, ob_space, ac_space, n_env, n_steps,
                                                    n_batch, transparent_params, reuse,
                                                    feature_extraction="mlp", **_kwargs)
-
-
-class TransparentCurryVecEnv(CurryVecEnv):
-    """CurryVecEnv that provides transparency data about its policy by updating infos dicts."""
-    def __init__(self, venv, policy, agent_idx=0, deterministic=False):
-        """
-        :param venv (VecMultiEnv): the environments
-        :param policy (BaseRLModel): model which wraps a BasePolicy object
-        :param agent_idx (int): the index of the agent that should be fixed.
-        :return: a new VecMultiEnv with num_agents decremented. It behaves like env but
-                 with all actions at index agent_idx set to those returned by agent."""
-        super().__init__(venv, policy, agent_idx, deterministic)
-        if not hasattr(self._policy.policy, 'step_transparent'):
-            raise TypeError("Error: policy must be transparent")
-        self._action = None
-
-    def step_async(self, actions):
-        policy_out = self._policy.predict_transparent(self._obs, state=self._state,
-                                                      mask=self._dones,
-                                                      deterministic=self.deterministic)
-        self._action, self._state, self._data = policy_out
-        actions.insert(self._agent_to_fix, self._action)
-        self.venv.step_async(actions)
-
-    def step_wait(self):
-        observations, rewards, self._dones, infos = self.venv.step_wait()
-        observations, self._obs = _tuple_pop(observations, self._agent_to_fix)
-        for env_idx in range(self.num_envs):
-            env_data = {k: v[env_idx] for k, v in self._data.items()}
-            infos[env_idx][self._agent_to_fix].update(env_data)
-        return observations, rewards, self._dones, infos
